test(api): migrate signup attribution to api-first bdd
Round 23 of the API-first BDD migration (#16967), SIGNUP ATTRIBUTION.

- Extend createBddApi with the attribution client and a
  mockClerkUserPrivateMetadata helper (Clerk owns the profile metadata, so the
  existing metadata is an external precondition).
- Add zero-attribution.bdd.test.ts: record first-touch attribution (mockNow for
  a deterministic recorded_at) and assert the merge preserves prior keys via the
  Clerk update mock; a second touch with existing signup_attribution reports
  recorded:false and writes nothing; plus a missing-session 401.
- Delete zero-attribution.test.ts whole.

Coverage verified (source-only): zero-attribution.ts (21/8/3) unchanged vs main.
No regressions.
